+++ vrepsim/objects.py
@@ -587,6 +587,35 @@ class VisionSensor(SceneObject):
 
         return image
 
+    def get_near_clip_plane(self, prec=VREP_FLOAT_PREC):
+        """Retrieve near clipping plane."""
+        if self._handle < 0:
+            if self._handle == MISSING_HANDLE:
+                raise RuntimeError(
+                    "Could not retrieve near clipping plane of {}: missing "
+                    "name or handle.".format(self._name))
+            if self._handle == REMOVED_OBJ_HANDLE:
+                raise RuntimeError(
+                    "Could not retrieve near clipping plane of {}: object "
+                    "removed.".format(self._name))
+        client_id = self.client_id
+        if client_id is None:
+            raise ConnectionError(
+                "Could not retrieve near clipping plane of {}: not connected "
+                "to V-REP remote API server.".format(self._name))
+        res, clip_plane = vrep.simxGetObjectFloatParameter(
+            client_id, self._handle, vrep.sim_visionfloatparam_near_clipping,
+            vrep.simx_opmode_blocking)
+        if res != vrep.simx_return_ok:
+            raise ServerError("Could not retrieve near clipping plane of {}."
+                              "".format(self._name))
+        if prec is not None:
+            clip_plane = round(clip_plane, prec)  # near clipping plane may be
+                                                  # slightly imprecise (about
+                                                  # the 10th digit after the
+                                                  # decimal point)
+        return clip_plane
+
 
 class MotorArray(object):
     """Interface to an array of motors simulated in V-REP."""